refactor: exclude the edit list key tests to a separate suite
re #RI-6570
